% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/dummyFunction.R
\name{dummyFunction}
\alias{dummyFunction}
\title{Dummy R function for nordummy}
\usage{
dummyFunction(doorToNeedleHours, moronScore)
}
\arguments{
\item{doorToNeedleHours}{Vector of hours from entry to treatment}

\item{moronScore}{Vector of relative stupidity}
}
\value{
density Density table of hours from entry to treatment

a histogram...
}
\description{
Dummy R function modtly to show how functions are documented.
Please make sure everything is documented in a way that others
can understand and use to maintain and alter the functions
}

